services

+++ DigitalBookStoreManagement/Service/BookManagementService.cs
@@ -1,4 +1,5 @@
-﻿using DigitalBookStoreManagement.Model;
+﻿using DigitalBookStoreManagement.Exception;
+using DigitalBookStoreManagement.Model;
 using DigitalBookStoreManagement.Repository;
 
 namespace DigitalBookStoreManagement.Service
@@ -15,12 +16,21 @@ namespace DigitalBookStoreManagement.Service
         public async Task<IEnumerable<BookManagement>> GetAllBooksAsync()
         {
             var books = await _repository.GetAllBooksAsync();
+            if (!books.Any())
+            {
+                throw new NotFoundException("No book records found.");
+            }
             return books;
         }
 
         public async Task<BookManagement?> GetBookByIdAsync(int bookId)
         {
-            return await _repository.GetBookByIdAsync(bookId);
+            var book = await _repository.GetBookByIdAsync(bookId);
+            if (book == null)
+            {
+                throw new NotFoundException($"Book with ID {bookId} not found.");
+            }
+            return book;
         }
 
         public async Task<IEnumerable<BookManagement>> SearchBooksByTitleAsync(string title)
@@ -40,16 +50,31 @@ namespace DigitalBookStoreManagement.Service
 
         public async Task<BookManagement> AddBookAsync(BookManagement book)
         {
+            var existingBook = await _repository.SearchBooksByTitleAsync(book.Title);
+            if (existingBook != null)
+            {
+                throw new AlreadyExistsException($"Book {book.Title} already exists.");
+            }
             return await _repository.AddBookAsync(book);
         }
 
         public async Task UpdateBookAsync(BookManagement book)
         {
+            var existingBook = await _repository.GetBookByIdAsync(book.BookID);
+            if (existingBook == null)
+            {
+                throw new NotFoundException($"Book with BookID {book.BookID} not found.");
+            }
             await _repository.UpdateBookAsync(book);
         }
 
         public async Task DeleteBookAsync(int bookId)
         {
+            var book = await _repository.GetBookByIdAsync(bookId);
+            if (book == null)
+            {
+                throw new NotFoundException($"Book with BookID {bookId} not found.");
+            }
             await _repository.DeleteBookAsync(bookId);
         }
     }
